feat/nodejs: Socket two way working
