+++ src/components/App.tsx
@@ -31,9 +31,9 @@ class App extends React.Component<IAppProps> {
         // THIS IS THE BASIC IPC - REDUX IMPLEMENTATION UNTIL A MIDDLEWARE IS
         // BUILD ON SERVERSIDE:
         let timer = setInterval(() => {
+            window.socketIoClient.emit('get-mixerprotocol', 'get selected mixerprotocol')
             window.socketIoClient.emit('get-store', 'update local store');
             window.socketIoClient.emit('get-settings', 'update local settings');
-            window.socketIoClient.emit('get-mixerprotocol', 'get selected mixerprotocol')
         },
         100)
